bug fix, small refactoring, better but not satisfying tests
