+++ attachments-verify/src/verify-sign/__tests__/wsaruba.test.ts
@@ -1,37 +1,13 @@
 import { right } from "fp-ts/lib/Either";
 import { task } from "fp-ts/lib/Task";
 import * as soap from "soap";
+import {
+  emailAttachmentsMock,
+  emailAttachmentsWithStatusMock
+} from "../../___mocks__/mocks";
 import { urlDemoAruba } from "../../domain/data";
 import { IEmailAttachmentStatus } from "../../domain/models";
 import * as ArubaVerify from "../../verify-sign/wsaruba";
-
-const emailAttachmentsMock = {
-  attachments: [
-    {
-      data: "base64string in binary format",
-      filename: "filename",
-      message: "64"
-    }
-  ],
-  date: ["Fri, 06 Dec 2019 22:38:57 +0000"],
-  from: ["ente@mail.com"],
-  messageId: "64",
-  subject: ["Registrazione presso la piattaforma IO"],
-  to: ["example@mail.com"]
-} as IEmailAttachmentStatus;
-
-const emailAttachmentsWithStatusMock = {
-  ...emailAttachmentsMock,
-  attachments: [
-    {
-      ...emailAttachmentsMock.attachments[0],
-      attachmentStatus: {
-        operation: "Verify PDF",
-        status: "OK"
-      }
-    }
-  ]
-} as IEmailAttachmentStatus;
 
 jest.mock("../../verify-sign/wsaruba", () => {
   const originalWsAruba = jest.requireActual("../../verify-sign/wsaruba");
